Cleng refactoring +++; CP(to_segments) is improved
