+++ src/cleng/cleng.cpp
@@ -1,6 +1,7 @@
 #include "cleng.h"
-#include "check_is_nan.h"
+#include "cleng_tools.h"
 //#include <unistd.h>
+#include "iterator/EnumerateIterator.h"
 
 using namespace std;
 
@@ -92,7 +93,7 @@ bool Cleng::CheckInput(int start, bool save_vector) {
 
         // delta_save
         if (!GetValue("delta_save").empty()) {
-            success = In[0]->Get_int(GetValue("delta_save"), delta_save, 1, MCS, "The delta_save interval should be between 1 and " + std::to_string(MCS));
+            success = In[0]->Get_int(GetValue("delta_save"), delta_save, 1, MCS, "The delta_save interval should be between 1 and " + to_string(MCS));
         } else delta_save = 1;
         if (debug) cout << "delta_save_interval " << delta_save << endl;
 
@@ -171,27 +172,6 @@ bool Cleng::CheckInput(int start, bool save_vector) {
     return success;
 }
 
-shared_ptr<SimpleNode> fromSystemToNode(int x, int y, int z, int id, const Point &box) {
-    return make_shared<SimpleNode>(Point(x, y, z), id, box);
-}
-
-vector<shared_ptr<Node>> createNodes(const vector<shared_ptr<SimpleNode>> &simple_nodes) {
-    vector<shared_ptr<Node>> result;
-    map<SimpleNode, vector<shared_ptr<SimpleNode>>> m;
-    for (auto &&n  : simple_nodes) {
-        m[*n].push_back(n);
-    }
-    for (auto &&entry : m) {
-        if (entry.second.size() == 1) {
-            result.push_back(entry.second[0]);
-        } else {
-            result.push_back(make_shared<Monolit>(entry.second));
-        }
-    }
-
-    return result;
-}
-
 bool Cleng::CP(transfer tofrom) {
     if (debug) cout << "CP in Cleng" << endl;
 
@@ -200,7 +180,7 @@ bool Cleng::CP(transfer tofrom) {
 
     int JX = Lat[0]->JX;
     int JY = Lat[0]->JY;
-    int M = Lat[0]->M;
+    int M  = Lat[0]->M;
 
     Segment *clamped = Seg[clamp_seg];
     map<int, Point> system_points;
@@ -208,83 +188,63 @@ bool Cleng::CP(transfer tofrom) {
         case to_cleng:
             simpleNodeList.clear();
             for (int i = 0; i < n_boxes; i++) {
-                auto first_node = fromSystemToNode(clamped->px1[i], clamped->py1[i], clamped->pz1[i], 2 * i, box);
+                auto first_node  = fromSystemToNode(clamped->px1[i], clamped->py1[i], clamped->pz1[i], 2 * i, box);
                 auto second_node = fromSystemToNode(clamped->px2[i], clamped->py2[i], clamped->pz2[i], 2 * i + 1, box);
-                first_node->set_cnode(second_node);
+                first_node ->set_cnode(second_node);
                 second_node->set_cnode(first_node);
                 //
                 simpleNodeList.push_back(first_node);
                 simpleNodeList.push_back(second_node);
-            }
-
-            nodes = createNodes(simpleNodeList);
-//            for (auto &&n : nodes) {
-//                cout << n->to_string() << endl;
-//            }
-//            assert(nodes.size() != 3);
+            }; nodes = createNodes(simpleNodeList);
             break;
 
         case to_segment:
             Zero(clamped->H_MASK, M);
 
-            for (auto &&n : nodes) n->pushSystemPoints(system_points);
+            for (auto &&SN : Enumerate(simpleNodeList)) {
+                size_t index = SN.first;    //
+                if (index % 2 == 1) continue;
+                size_t n_box = index / 2;   // n_boxes
+                Point p1 = SN.second->point();                // all time in box [0:box_l]
+                Point p2 = SN.second->get_cnode()->point();   // all time in box [0:box_l]
 
-            for (int index=0; index < (int) system_points.size(); index=index+2) {
-                auto first_node = system_points[index];
-                auto second_node = system_points[index+1];
+                Point p1sys  = SN.second->get_system_point();
+                Point p2sys  = SN.second->get_cnode()->get_system_point();
 
-                if (!first_node.point_in_range(box) and !second_node.point_in_range(box)) {
-                    cout << "1* box" << endl;
-                    system_points[index]   = first_node % box;
-                    system_points[index+1] = second_node % box;
-                }
-                if (!first_node.point_in_range(box+box) and !second_node.point_in_range(box+box)) {
-                    cout << "2* box" << endl;
-                    system_points[index]   = first_node % box;
-                    system_points[index+1] = second_node % box;
-                }
-            }
+//                cout << " n_box: " << n_box << endl;
+//                cout << " index: " << index << " v: " << p1.x << " " << "("<<p1sys.x<<")" << " " << p1.y << " "  << "("<<p1sys.y<<")" << " " << p1.z << " "  << "("<<p1sys.z<<")" << endl;
+//                cout << " index: " << index << " v: " << p2.x << " " << "("<<p2sys.x<<")" << " " << p2.y << " "  << "("<<p2sys.y<<")" << " " << p2.z << " "  << "("<<p2sys.z<<")" << endl;
 
-            for (auto &&entry : system_points) {
-                int i = entry.first / 2;
-                Point &p = entry.second;
-                if (entry.first % 2 == 0) {
-                    clamped->px1[i] = p.x;
-                    clamped->py1[i] = p.y;
-                    clamped->pz1[i] = p.z;
-                } else {
-                    clamped->px2[i] = p.x;
-                    clamped->py2[i] = p.y;
-                    clamped->pz2[i] = p.z;
-                }
-            }
+                if ((p1.x - p2.x) != (p1sys.x - p2sys.x)) {if (p1.x < p2.x) p1.x += box.x; else p2.x += box.x;}
+                if ((p1.y - p2.y) != (p1sys.y - p2sys.y)) {if (p1.y < p2.y) p1.y += box.y; else p2.y += box.y;}
+                if ((p1.z - p2.z) != (p1sys.z - p2sys.z)) {if (p1.z < p2.z) p1.z += box.z; else p2.z += box.z;}
 
-            for (int i = 0; i < n_boxes; i++) {
-                clamped->bx[i] = (clamped->px2[i] + clamped->px1[i] - sub_box_size) / 2;
-                clamped->by[i] = (clamped->py2[i] + clamped->py1[i] - sub_box_size) / 2;
-                clamped->bz[i] = (clamped->pz2[i] + clamped->pz1[i] - sub_box_size) / 2;
+                clamped->px1[n_box] = p1.x;
+                clamped->py1[n_box] = p1.y;
+                clamped->pz1[n_box] = p1.z;
 
-                if (clamped->bx[i] < 1) {
-                    clamped->bx[i] += box.x;
-                    clamped->px1[i] += box.x;
-                    clamped->px2[i] += box.x;
-                }
-                if (clamped->by[i] < 1) {
-                    clamped->by[i] += box.y;
-                    clamped->py1[i] += box.y;
-                    clamped->py2[i] += box.y;
-                }
-                if (clamped->bz[i] < 1) {
-                    clamped->bz[i] += box.z;
-                    clamped->pz1[i] += box.z;
-                    clamped->pz2[i] += box.z;
-                }
+                clamped->px2[n_box] = p2.x;
+                clamped->py2[n_box] = p2.y;
+                clamped->pz2[n_box] = p2.z;
+// box
+                clamped->bx[n_box] = (p2.x + p1.x - sub_box_size) / 2;
+                clamped->by[n_box] = (p2.y + p1.y - sub_box_size) / 2;
+                clamped->bz[n_box] = (p2.z + p1.z - sub_box_size) / 2;
 
-                clamped->H_MASK[((clamped->px1[i] - 1) % box.x + 1) * JX + ((clamped->py1[i] - 1) % box.y + 1) * JY +
-                                (clamped->pz1[i] - 1) % box.z + 1] = 1;
-                clamped->H_MASK[((clamped->px2[i] - 1) % box.x + 1) * JX + ((clamped->py2[i] - 1) % box.y + 1) * JY +
-                                (clamped->pz2[i] - 1) % box.z + 1] = 1;
+                if (clamped->bx[n_box] < 1) { clamped->bx[n_box] += box.x; clamped->px1[n_box] += box.x; clamped->px2[n_box] += box.x;}
+                if (clamped->by[n_box] < 1) { clamped->by[n_box] += box.y; clamped->py1[n_box] += box.y; clamped->py2[n_box] += box.y;}
+                if (clamped->bz[n_box] < 1) { clamped->bz[n_box] += box.z; clamped->pz1[n_box] += box.z; clamped->pz2[n_box] += box.z;}
+// clearing
+                auto hp1x = ((clamped->px1[n_box] - 1) % box.x + 1) * JX;
+                auto hp1y = ((clamped->py1[n_box] - 1) % box.y + 1) * JY;
+                auto hp1z =  (clamped->pz1[n_box] - 1) % box.z + 1;
 
+                auto hp2x = ((clamped->px2[n_box] - 1) % box.x + 1) * JX;
+                auto hp2y = ((clamped->py2[n_box] - 1) % box.y + 1) * JY;
+                auto hp2z = ( clamped->pz2[n_box] - 1) % box.z + 1;
+
+                clamped->H_MASK[hp1x + hp1y + hp1z] = 1;
+                clamped->H_MASK[hp2x + hp2y + hp2z] = 1;
             }
             break;
 
@@ -294,81 +254,6 @@ bool Cleng::CP(transfer tofrom) {
             break;
     }
     return success;
-}
-
-Real Cleng::GetN_times_mu() {
-    int n_mol = (int) In[0]->MolList.size();
-    Real n_times_mu = 0;
-    for (int i = 0; i < n_mol; i++) {
-        Real Mu = Mol[i]->Mu;
-        Real n = Mol[i]->n;
-        if (Mol[i]->IsClamped()) n = Mol[i]->n_box;
-        n_times_mu += n * Mu;
-    }
-    return n_times_mu;
-}
-
-bool Cleng::InSubBoxRange() {
-//    cout << "InSubBoxRange ... " << endl;
-    int not_in_subbox_range = 0;
-    Point sub_box = {sub_box_size-2, sub_box_size-2, sub_box_size-2};  // change it in future
-//    cout << "sub_box: " << sub_box.to_string() << endl;
-    if (!nodes[id_node_for_move]->inSubBoxRange(sub_box, clamped_move)) not_in_subbox_range++;
-
-    if (not_in_subbox_range != 0) {
-        cout << "There are nodes not in sub-box range!" << endl;
-        return false;
-    }
-    return true;
-}
-
-bool Cleng::NotCollapsing() {
-    bool not_collapsing = false;
-    Point MP(nodes[id_node_for_move]->point());
-    Point MPs(nodes[id_node_for_move]->point() + clamped_move);
-
-//    cout << "MP: "  << MP.to_string() << endl;
-//    cout << "MPs: " << MPs.to_string() << endl;
-
-    double min_dist = 2; // minimal distance between nodes
-    int i = 0;
-    for (const auto &n : nodes) {
-        if (MP != n->point()) {
-            Real distance = MPs.distance(n->point());
-            if (distance < min_dist) {
-                cout << "Nodes are too close to each other." << endl;
-                i++;
-            }
-        }
-    }
-    if (i == 0) {
-        not_collapsing = true;
-    }
-    return not_collapsing;
-}
-
-bool Cleng::InRange() {
-    bool in_range = false;
-    Point box = {Lat[0]->MX, Lat[0]->MY, Lat[0]->MZ};
-    Point down_boundary = {1, 1, 1};
-    Point up_boundary = box - down_boundary;
-    Point MPs(nodes[id_node_for_move]->point() + clamped_move);
-
-//    cout << "MPs" << MPs.to_string() << endl;
-//    cout << "down bound" << down_boundary.to_string() << endl;
-//    cout << "up boundary bound" << up_boundary.to_string() << endl;
-
-    if ((down_boundary.all_elements_less_than(MPs)) and (MPs.all_elements_less_than(up_boundary))) in_range = true;
-    return in_range;
-}
-
-void Cleng::make_BC() {
-//    make boundary condition point => BC = {1,1,1} => minor; BC = {0,0,0} => periodic
-    BC = {
-            Lat[0]->BC[0] == "mirror",
-            Lat[0]->BC[2] == "mirror",
-            Lat[0]->BC[4] == "mirror",
-    };
 }
 
 bool Cleng::Checks() {
@@ -391,56 +276,6 @@ bool Cleng::Checks() {
 
     result = not_collapsing and in_range and in_subbox_range;
     return result;
-}
-
-vector<int> makeExcludedArray(int step) {
-    vector<int> result;
-    for (int i = -step + 1; i < step; i++) result.push_back(i);
-    return result;
-}
-
-Point Cleng::prepareMove() {
-
-    if (debug) cout << "prepareMove in Cleng" << endl;
-
-    if (axis != -1) {
-        if (axis == 1) clamped_move = {-2, 0, 0};
-        if (axis == 2) clamped_move = {0, -2, 0};
-        if (axis == 3) clamped_move = {0, 0, -2};
-    } else {
-
-        clamped_move = {
-                rand.getIntExcludeArray(-delta_step, delta_step, makeExcludedArray(delta_step)),
-                rand.getIntExcludeArray(-delta_step, delta_step, makeExcludedArray(delta_step)),
-                rand.getIntExcludeArray(-delta_step, delta_step, makeExcludedArray(delta_step)),
-        };
-
-        if ((delta_step % 2) == 1) {  // 1 3 5
-            int id4rm = rand.getInt(0, 2);
-            if (id4rm == 0) clamped_move.x = 0;
-            else {
-                if (id4rm == 1) clamped_move.y = 0;
-                else clamped_move.z = 0;
-            }
-        } else {  // 2 4 6
-            int id4rm1 = rand.getInt(0, 2);
-            int id4rm2 = rand.getInt(0, 2);
-
-            if (id4rm1 == 0) clamped_move.x = 0;
-            else {
-                if (id4rm1 == 1) clamped_move.y = 0;
-                else clamped_move.z = 0;
-            }
-
-            if (id4rm2 == 0) clamped_move.x = 0;
-            else {
-                if (id4rm2 == 1) clamped_move.y = 0;
-                else clamped_move.z = 0;
-            }
-        }
-    }
-
-    return clamped_move;
 }
 
 bool Cleng::MakeMove(bool back) {
@@ -490,34 +325,6 @@ bool Cleng::MakeMove(bool back) {
     return success;
 }
 
-
-int Cleng::getLastMCS() {
-
-    string filename;
-    vector<string> sub;
-    int MS_step = 0;
-
-    string infilename = In[0]->name;
-    In[0]->split(infilename, '.', sub);
-    filename = sub[0];
-    filename = In[0]->output_info.getOutputPath() + filename;
-    //read kal file
-    ifstream infile(filename + ".kal");
-
-    if (infile) {
-        string line;
-        while (infile >> std::ws && std::getline(infile, line)); // skip empty lines
-
-        std::istringstream iss(line);
-        std::vector<std::string> results(std::istream_iterator<std::string>{iss},
-                                         std::istream_iterator<std::string>());
-
-        MS_step = stoi(results[0]);
-    } else cout << "Unable to open kal file.\n";
-
-    return MS_step;
-}
-
 bool Cleng::MonteCarlo(bool save_vector) {
     if (debug) cout << "Monte Carlo in Cleng" << endl;
     bool success = true;
@@ -556,12 +363,11 @@ bool Cleng::MonteCarlo(bool save_vector) {
         CP(to_segment);
 
         cout << endl;
-        cout << "System for calculation: " << endl;
+        cout << "[Cleng] System for calculation: " << endl;
         for (auto &&n : nodes) cout << n->to_string() << endl;
         cout << endl;
 
         if (success_) {
-
             New[0]->Solve(true);
             free_energy_trial = Sys[0]->GetFreeEnergy() - GetN_times_mu();
             if (save_vector) test_vector.push_back(free_energy_trial);
@@ -569,10 +375,7 @@ bool Cleng::MonteCarlo(bool save_vector) {
             cout << "Free Energy (c): " << free_energy_current    << endl;
             cout << "            (t): " << free_energy_trial << endl;
 //            cout << "trial - current = " << std::to_string(free_energy_trial - free_energy_current) << endl;
-            if (is_ieee754_nan( free_energy_trial )) {
-                cout << " Sorry, Free Energy is NaN. Termination..." << endl;
-                break;
-            }
+            if (is_ieee754_nan( free_energy_trial )) {cout << " Sorry, Free Energy is NaN. Termination..." << endl; break;}
 
             if (free_energy_trial - free_energy_current <= 0.0) {
                 cout << "Accepted" << endl;
@@ -587,7 +390,6 @@ bool Cleng::MonteCarlo(bool save_vector) {
                     accepted++;
                 } else {
                     cout << "Rejected" << endl;
-                    CP(to_cleng);
                     MakeMove(true);
                     CP(to_segment);
                     rejected++;
@@ -598,9 +400,7 @@ bool Cleng::MonteCarlo(bool save_vector) {
         cout << "Accepted: # " << accepted << " | " << 100 * (accepted / MC_attempt) << "%" << endl;
         cout << "Rejected: # " << rejected << " | " << 100 * (rejected / MC_attempt) << "%" << endl;
 
-        if ((MC_attempt % delta_save) == 0) {
-            WriteOutput(MC_attempt + MCS_checkpoint);
-        }
+        if ((MC_attempt % delta_save) == 0) WriteOutput(MC_attempt + MCS_checkpoint);
         if (checkpoint_save) checkpoint.updateCheckpoint(simpleNodeList);
     }
 
@@ -610,117 +410,7 @@ bool Cleng::MonteCarlo(bool save_vector) {
     cout << "Accepted: # " << accepted << " | " << 100 * (accepted / (MC_attempt-1)) << "%" << endl;
     cout << "Rejected: # " << rejected << " | " << 100 * (rejected / (MC_attempt-1)) << "%" << endl;
 
-    if (checkpoint_save) {
-        cout << "Saving checkpoint..." << endl;
-        checkpoint.saveCheckpoint(simpleNodeList);
-    }
-
+    if (checkpoint_save) { cout << "Saving checkpoint..." << endl; checkpoint.saveCheckpoint(simpleNodeList);}
     cout << "Have a fun. " << endl;
-
     return success;
-}
-
-void Cleng::WriteOutput(int attempt) {
-    if (debug) cout << "WriteOutput in Cleng" << endl;
-    PushOutput(attempt);
-    New[0]->PushOutput();
-    for (int i = 0; i < n_out; i++) Out[i]->WriteOutput(attempt);
-    if (cleng_pos) WriteClampedNodeDistance(attempt);
-}
-
-void Cleng::PushOutput(int attempt) {
-//	int* point;
-    for (int i = 0; i < n_out; i++) {
-        Out[i]->PointerVectorInt.clear();
-        Out[i]->PointerVectorReal.clear();
-        Out[i]->SizeVectorInt.clear();
-        Out[i]->SizeVectorReal.clear();
-        Out[i]->strings.clear();
-        Out[i]->strings_value.clear();
-        Out[i]->bools.clear();
-        Out[i]->bools_value.clear();
-        Out[i]->Reals.clear();
-        Out[i]->Reals_value.clear();
-        Out[i]->ints.clear();
-        Out[i]->ints_value.clear();
-
-        if (Out[i]->name == "ana" || Out[i]->name == "kal") {
-            Out[i]->push("MC_attempt", attempt);
-            Out[i]->push("MCS", MCS);
-            Out[i]->push("free_energy", free_energy_current);
-
-        }
-
-        if (Out[i]->name == "ana" ||
-            Out[i]->name == "vec") { //example for putting an array of Reals of arbitrary length to output
-            string s = "vector;0"; //the keyword 'vector' is used for Reals; the value 0 is the first vector, use 1 for the next etc,
-            Out[i]->push("gn", s); //"gn" is the name that will appear in output file
-            Out[i]->PointerVectorReal.push_back(
-                    Mol[clp_mol]->gn); //this is the pointer to the start of the 'vector' that is reported to output.
-            Out[i]->SizeVectorReal.push_back(
-                    sizeof(Mol[clp_mol]->gn)); //this is the size of the 'vector' that is reported to output
-        }
-
-        if (Out[i]->name == "ana" ||
-            Out[i]->name == "pos") { //example for putting 3 array's of Integers of arbitrary length to output
-            string s = "array;0";
-            Out[i]->push("X", s);
-
-//			 TODO using normal point in output and remove xs, ys, zs
-            fillXYZ();
-//			point=X.data();
-            Out[i]->PointerVectorInt.push_back(xs);
-            Out[i]->SizeVectorInt.push_back((int) nodes.size());
-            s = "array;1";
-            Out[i]->push("Y", s);
-//			point = Y.data();
-            Out[i]->PointerVectorInt.push_back(ys);
-            Out[i]->SizeVectorInt.push_back((int) nodes.size());
-            s = "array;2";
-            Out[i]->push("Z", s);
-//			point=Z.data();
-            Out[i]->PointerVectorInt.push_back(zs);
-            Out[i]->SizeVectorInt.push_back((int) nodes.size());
-        }
-    }
-}
-
-void Cleng::fillXYZ() {
-    delete[] xs;
-    delete[] ys;
-    delete[] zs;
-
-    int n = nodes.size();
-    xs = new int[n];
-    ys = new int[n];
-    zs = new int[n];
-    for (int i = 0; i < n; i++) {
-        xs[i] = nodes[i]->point().x;
-        ys[i] = nodes[i]->point().y;
-        zs[i] = nodes[i]->point().z;
-    }
-}
-
-void Cleng::WriteClampedNodeDistance(int MS_step) {
-    vector<Real> distPerMC;
-    int i = 0;
-    for (auto &&SN :  simpleNodeList) {
-        if (!(i % 2)) distPerMC.push_back(SN->distance(SN->get_cnode()->get_system_point()));
-        i++;
-    }
-
-    string filename;
-    vector<string> sub;
-
-    string infilename = In[0]->name;
-    In[0]->split(infilename, '.', sub);
-    filename = sub[0];
-    filename = In[0]->output_info.getOutputPath() + filename;
-
-    ofstream outfile;
-    outfile.open(filename + ".cpos", std::ios_base::app);
-
-    outfile << MS_step << " ";
-    for (auto n : distPerMC)outfile << n << " ";
-    outfile << endl;
 }
